backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.17
SQLAlchemy==2.0.32
python-dotenv==1.0.1
psycopg2==2.9.9
redis==5.2.0
bleach==6.2.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
